requirements.txt
streamlit
biopython
numpy
plotly
matplotlib
typing-extensions
kaleido
# Install pyfamsa from a wheel file for better compatibility on hosted servers
pyfamsa==0.3.2
pandas
scikit-learn
